Major performance fix: 3x GA time + float angle + active blocker
1. ELIMINATED GA OPPONENT MODEL (was stealing 45ms/70ms budget)
   - Replaced expensive RunGA opponent prediction with zero-cost heuristic
   - Our GA now gets ~65ms instead of 25ms (2.6x more planning time)
   - Runner will no longer output absurd thrust=0/6/22 values

2. FLOAT ANGLE PRECISION (eliminates ALL PHYSICS BUG entries)
   - Changed Pod::angle from int to double across all files
   - Magus referee stores angle as float, accumulating fractional diffs
   - Our int truncation was causing 1° drift every turn
   - Verifier now uses precise trig (cos/sin of double angle)

3. RESTORED BLOCKER EFFECTIVENESS
   - block_weight: 0.1 -> 5.0 (blocker actually blocks now)
   - opp_penalty: 0.0 -> 1.0 (penalize opponent progress)
   - HeuristicBlocker (already in use) now has meaningful weight

Synced changes across cg_bot.cpp, engine.h, engine.cpp, ga_bot.cpp

+++ src/engine/engine.h
@@ -45,7 +45,7 @@ struct Pod {
     int id;
     int team;
     Vec2 pos, vel;
-    int angle;
+    double angle;
     int next_cp_id;
     bool boost_available;
     int shield_cd;
